+++ src/hall_sensor_node.cpp
@@ -39,8 +39,8 @@ int main(int argc, char** argv)
     uint8_t spi_cs_id = static_cast<uint8_t>(spi_cs_id_int);
     uint8_t spi_mode = static_cast<uint8_t>(spi_mode_int);
     uint8_t spi_bits = static_cast<uint8_t>(spi_bits_int);
-    uint8_t spi_speed = static_cast<uint8_t>(spi_speed_int);
-    uint8_t spi_delay = static_cast<uint8_t>(spi_delay_int);
+    uint32_t spi_speed = static_cast<uint32_t>(spi_speed_int);
+    uint16_t spi_delay = static_cast<uint16_t>(spi_delay_int);
 
     ROS_DEBUG("Node: device: %s, id: %u, mode: % u, bits: % u, speed: % u, delay: % u \n", spi_device, spi_cs_id, spi_mode, spi_bits, spi_speed, spi_delay);
